Creating a basic usb printer class
This change shuffles some functionality out of the usbip header, so
that it serves as more of a utility rather than a main program.

A new server program is being introduced to replace the usbip_run()
function.

This change also modifies the code to use C++ in order to take advantage
of some more modern features and abstractions.